making mypy compliant code

+++ src/congkak/game/custom_scenarios.py
@@ -23,10 +23,10 @@ CUSTOM_GAME_INDEX_TO_BOARD_STATE_MAPPING = {
         player_one=Player(
             number=PlayerNumber.ONE,
             score=0,
-            side=np.zeros(PITS_PER_SIDE),
+            side=np.zeros(PITS_PER_SIDE, dtype=np.int32),
         ),
         player_two=Player(
-            number=PlayerNumber.TWO, score=0, side=np.zeros(PITS_PER_SIDE)
+            number=PlayerNumber.TWO, score=0, side=np.zeros(PITS_PER_SIDE, dtype=np.int32)
         ),
     ),
     CustomGameScenario.FREE_GO: BoardState(
@@ -35,10 +35,10 @@ CUSTOM_GAME_INDEX_TO_BOARD_STATE_MAPPING = {
         player_one=Player(
             number=PlayerNumber.ONE,
             score=0,
-            side=np.array([0 for _ in range(PITS_PER_SIDE - 1)] + [1]),
+            side=np.array([0 for _ in range(PITS_PER_SIDE - 1)] + [1], dtype=np.int32),
         ),
         player_two=Player(
-            number=PlayerNumber.TWO, score=0, side=np.zeros(PITS_PER_SIDE)
+            number=PlayerNumber.TWO, score=0, side=np.zeros(PITS_PER_SIDE, dtype=np.int32)
         ),
     ),
     CustomGameScenario.STEAL: BoardState(
@@ -47,12 +47,12 @@ CUSTOM_GAME_INDEX_TO_BOARD_STATE_MAPPING = {
         player_one=Player(
             number=PlayerNumber.ONE,
             score=0,
-            side=np.array([1] + [0 for _ in range(PITS_PER_SIDE - 1)]),
+            side=np.array([1] + [0 for _ in range(PITS_PER_SIDE - 1)], dtype=np.int32),
         ),
         player_two=Player(
             number=PlayerNumber.TWO,
             score=0,
-            side=np.array([0] + [0 for _ in range(PITS_PER_SIDE - 2)] + [0]),
+            side=np.array([0] + [0 for _ in range(PITS_PER_SIDE - 2)] + [0], dtype=np.int32),
         ),
     ),
     CustomGameScenario.LOSE_GO: BoardState(
@@ -61,10 +61,10 @@ CUSTOM_GAME_INDEX_TO_BOARD_STATE_MAPPING = {
         player_one=Player(
             number=PlayerNumber.ONE,
             score=0,
-            side=np.array([0 for _ in range(PITS_PER_SIDE - 1)] + [2]),
+            side=np.array([0 for _ in range(PITS_PER_SIDE - 1)] + [2], dtype=np.int32),
         ),
         player_two=Player(
-            number=PlayerNumber.TWO, score=0, side=np.zeros(PITS_PER_SIDE)
+            number=PlayerNumber.TWO, score=0, side=np.zeros(PITS_PER_SIDE, dtype=np.int32)
         ),
     ),
     CustomGameScenario.SCENARIO_1: BoardState(
@@ -73,10 +73,10 @@ CUSTOM_GAME_INDEX_TO_BOARD_STATE_MAPPING = {
         player_one=Player(
             number=PlayerNumber.ONE,
             score=0,
-            side=np.ones(PITS_PER_SIDE),
+            side=np.ones(PITS_PER_SIDE, dtype=np.int32),
         ),
         player_two=Player(
-            number=PlayerNumber.TWO, score=0, side=np.ones(PITS_PER_SIDE)
+            number=PlayerNumber.TWO, score=0, side=np.ones(PITS_PER_SIDE, dtype=np.int32)
         ),
     ),
 }
